post

+++ app/Http/Controllers/TasksController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 
 use App\Task;
 
+
+
 class TasksController extends Controller
 {
     /**
@@ -15,12 +17,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //
-        $tasks = Task::all();
-        
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
 
     /**
@@ -57,8 +68,10 @@ class TasksController extends Controller
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
-        $task->save();
+        
+       $task->user_id = \Auth::id();
 
+        $task->save();
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -134,9 +147,10 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
-
+        
+           if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
